Google Nav Bar, Redesign Home UI, Fix Minor Bug

+++ lib/services/upload_image_service.dart
@@ -15,7 +15,8 @@ class UploadImageService {
     return "";
   }
 
-  static void deleteImage(String email) async {
+  static void deleteImage(String email, File image) async {
+    if (image == null) return;
     await _userProfileRef.child("$email.png").delete();
   }
 }
